+++ src/Clients/EnableU/Authenticator.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OWC\ZGW\Clients\EnableU;
+
+use OWC\ZGW\Contracts\AbstractTokenAuthenticator;
+
+class Authenticator extends AbstractTokenAuthenticator
+{
+}