+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.0] - 2026-07-29
+
+### Added
+
+- Support for the EnableU client
+- Support for 'zaakverzoeken' endpoint
+
+### Fixed
+
+- Loads and loads of issues flagged by PHPStan
+
 ## [2.5.3] - 2026-06-23
 
 ### Fixed